requirements.txt
Flask
redis
python-dotenv
sendgrid
twilio
azure-communication-email
